Push index expansion logging into base visitor, extending classes provide stage name
